26/04/2023 - Update 005
modified entity stats to use IDamageable, had to modify the firearm & melee scripts as they used floats, but damage should be in int.

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -6,6 +6,6 @@ namespace com.limphus.retro_survival_shooter
 {
     public interface IDamageable
     {
-        void Damage(float amount);
+        void Damage(int amount);
     }
 }
